requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.2
Django==3.2.11
django-celery-beat==2.2.1
django-timezone-field==4.2.1
djangorestframework==3.13.1
flower==1.0.0
humanize==3.13.1
kombu==5.2.3
prometheus-client==0.12.0
prompt-toolkit==3.0.24
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
